move file and update mdx components
