+++ docs/components/mdxComponents.tsx
@@ -15,7 +15,7 @@ import generatedComponentPropsData from '../__generated__/componentProps.json';
 import { Code } from './Code';
 import { ComponentPropsTable } from './ComponentPropsTable';
 import { FigmaLogo } from './FigmaLogo';
-import { DoHeading, DontHeading } from './DoHeading';
+import { DoHeading, DontHeading } from './DoDontHeading';
 
 export const mdxComponents: MDXRemoteProps['components'] = {
 	Fragment,
@@ -77,6 +77,6 @@ export const mdxComponents: MDXRemoteProps['components'] = {
 			/>
 		);
 	},
-	DoHeading: () => <DoHeading />,
-	DontHeading: () => <DontHeading />,
+	DoHeading,
+	DontHeading,
 };
